refactor: use boost/core/ref.hpp over boost/ref.hpp
boost/ref.hpp is deprecated.
