+++ include/boost/signals2/detail/signals_common.hpp
@@ -11,9 +11,9 @@
 #ifndef BOOST_SIGNALS2_SIGNALS_COMMON_HPP
 #define BOOST_SIGNALS2_SIGNALS_COMMON_HPP
 
+#include <boost/core/ref.hpp>
 #include <boost/mpl/bool.hpp>
 #include <boost/mpl/if.hpp>
-#include <boost/ref.hpp>
 #include <boost/signals2/signal_base.hpp>
 #include <boost/type_traits/is_base_of.hpp>
